Update Current Baseline
This Push has changed the realtime_monitor.py so that it has a system where it displays the changes from the baseline to current. And prompt the user if they wanna update the Baseline to the recently created snapshot. And in 20 seconds if the user decides not to do anything it will change nothing and continue its monitoring.