Added CloudTail to see IP of changes in drift
CloudTrail logs from AWS taken and logged into the Realtime_monitor.py to be output into the log. It will show you the changes from the drift and where it comes from and who and what exaclty was changed. If the snapshot is similar than the one 20 seconds ago, then it will not show any changes on CloudTrail.